+++ ntex-rt/src/system.rs
@@ -1,12 +1,22 @@
 use std::sync::{atomic::AtomicUsize, atomic::Ordering, Arc};
-use std::{cell::RefCell, fmt, future::Future, pin::Pin, rc::Rc};
+use std::{cell::RefCell, collections::HashMap, fmt, future::Future, pin::Pin, rc::Rc};
 
-use async_channel::Sender;
+use async_channel::{Receiver, Sender};
 
-use super::arbiter::{Arbiter, SystemCommand};
+use super::arbiter::Arbiter;
 use super::builder::{Builder, SystemRunner};
 
 static SYSTEM_COUNT: AtomicUsize = AtomicUsize::new(0);
+
+thread_local!(
+    static ARBITERS: RefCell<Arbiters> = RefCell::new(Arbiters::default());
+);
+
+#[derive(Default)]
+struct Arbiters {
+    all: HashMap<usize, Arbiter>,
+    list: Vec<Arbiter>,
+}
 
 /// System is a runtime manager.
 #[derive(Clone, Debug)]
@@ -33,14 +43,17 @@ impl System {
     /// Constructs new system and sets it as current
     pub(super) fn construct(
         sys: Sender<SystemCommand>,
-        arbiter: Arbiter,
+        mut arbiter: Arbiter,
         config: SystemConfig,
     ) -> Self {
+        let id = SYSTEM_COUNT.fetch_add(1, Ordering::SeqCst);
+        arbiter.sys_id = id;
+
         let sys = System {
+            id,
             sys,
             config,
             arbiter,
-            id: SYSTEM_COUNT.fetch_add(1, Ordering::SeqCst),
         };
         System::set_current(sys.clone());
         sys
@@ -104,6 +117,17 @@ impl System {
         &self.arbiter
     }
 
+    /// Retrieves a list of all arbiters in the system.
+    ///
+    /// This method should be called from the context where the system has been initialized
+    pub fn list_arbiters<F, R>(f: F) -> R
+    where F: FnOnce(&[Arbiter]) -> R,
+    {
+        ARBITERS.with(|arbs| {
+            f(arbs.borrow().list.as_ref())
+        })
+    }
+
     pub(super) fn sys(&self) -> &Sender<SystemCommand> {
         &self.sys
     }
@@ -148,5 +172,100 @@ impl fmt::Debug for SystemConfig {
             .field("stack_size", &self.stack_size)
             .field("stop_on_panic", &self.stop_on_panic)
             .finish()
+    }
+}
+
+#[derive(Debug)]
+pub(super) enum SystemCommand {
+    Exit(i32),
+    RegisterArbiter(usize, Arbiter),
+    UnregisterArbiter(usize),
+}
+
+pub(super) struct SystemSupport {
+    stop: Option<oneshot::Sender<i32>>,
+    commands: Receiver<SystemCommand>,
+    ping_interval: usize,
+}
+
+impl SystemSupport {
+    pub(super) fn new(
+        stop: oneshot::Sender<i32>,
+        commands: Receiver<SystemCommand>,
+        ping_interval: usize,
+    ) -> Self {
+        Self {
+            commands,
+            ping_interval,
+            stop: Some(stop),
+        }
+    }
+
+    pub(super) async fn run(mut self) {
+        ARBITERS.with(move |arbs| {
+            let mut arbiters = arbs.borrow_mut();
+            arbiters.all.clear();
+            arbiters.list.clear();
+        });
+
+        loop {
+            match self.commands.recv().await {
+                Ok(SystemCommand::Exit(code)) => {
+                    log::debug!("Stopping system with {} code", code);
+
+                    // stop arbiters
+                    ARBITERS.with(move |arbs| {
+                        let mut arbiters = arbs.borrow_mut();
+                        for arb in arbiters.list.drain(..) {
+                            arb.stop();
+                        }
+                        arbiters.all.clear();
+                    });
+
+                    // stop event loop
+                    if let Some(stop) = self.stop.take() {
+                        let _ = stop.send(code);
+                    }
+                }
+                Ok(SystemCommand::RegisterArbiter(name, hnd)) => {
+                    ARBITERS.with(move |arbs| {
+                        let mut arbiters = arbs.borrow_mut();
+                        arbiters.all.insert(name, hnd.clone());
+                        arbiters.list.push(hnd);
+                    })
+                }
+                Ok(SystemCommand::UnregisterArbiter(id)) => {
+                    ARBITERS.with(move |arbs| {
+                        let mut arbiters = arbs.borrow_mut();
+                        if let Some(hnd) = arbiters.all.remove(&id) {
+                            for (idx, arb) in arbiters.list.iter().enumerate() {
+                                if &hnd == arb {
+                                    arbiters.list.remove(idx);
+                                    break
+                                }
+                            }
+                        }
+                    });
+                }
+                Err(_) => {
+                    log::debug!("System stopped");
+                    return;
+                }
+            }
+        }
+    }
+}
+
+pub(super) trait FnExec: Send + 'static {
+    fn call_box(self: Box<Self>);
+}
+
+impl<F> FnExec for F
+where
+    F: FnOnce() + Send + 'static,
+{
+    #[allow(clippy::boxed_local)]
+    fn call_box(self: Box<Self>) {
+        (*self)()
     }
 }